% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agglomUA-package.R
\docType{package}
\name{agglomUA-package}
\alias{agglomUA}
\alias{agglomUA-package}
\title{agglomUA: Combine US Census Bureau Uraban Areas}
\description{
An R package containing functions to approximate US Census Bureau "larger urban agglomerations".
}
\author{
\strong{Maintainer}: Aaron Dickey \email{adickey7@hotmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
