% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_summaries.R
\name{checksf}
\alias{checksf}
\title{check the coordinate reference system of a sf}
\usage{
checksf(ua)
}
\arguments{
\item{ua}{a sf}
}
\value{
an error if the sf lacks a coordinate reference system
}
\description{
check the coordinate reference system of a sf
}
\examples{
data("ua2000"); class(ua2000)
#will produce error (not run) `checksf(ua2000)`
}
