% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_summaries.R
\name{checkuasynonymy}
\alias{checkuasynonymy}
\title{check for urban area name and uace code matching between urban area sf and an urban area metadata data frame}
\usage{
checkuasynonymy(ua, ua_meta)
}
\arguments{
\item{ua}{a sf}

\item{ua_meta}{a data frame}
}
\value{
an error if a synonymy error is found
}
\description{
check for urban area name and uace code matching between urban area sf and an urban area metadata data frame
}
\examples{
data("ua2000"); data("meta2000"); class(ua2000); class(meta2000)
#will produce error (not run) `checkuasynonymy(ua2000, meta2000)`
}
