% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_summaries.R
\name{decimalplaces}
\alias{decimalplaces}
\title{return the number of decimal places}
\usage{
decimalplaces(x)
}
\arguments{
\item{x}{a number}
}
\value{
an integer
}
\description{
return the number of decimal places
}
\examples{
decimalplaces(5.678)
}
