% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_summaries.R
\name{gmUA}
\alias{gmUA}
\title{apply a gravitational model to a unioned (see ?unionUA) urban area sf}
\usage{
gmUA(ua)
}
\arguments{
\item{ua}{a unioned sf}
}
\value{
a sf
}
\description{
apply a gravitational model to a unioned (see ?unionUA) urban area sf
}
\examples{
#none
}
