% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_summaries.R
\name{mergeUA}
\alias{mergeUA}
\title{Merge a compatible urban area sf with its matching metadata data frame}
\usage{
mergeUA(wd, file, mfile)
}
\arguments{
\item{wd}{a directory containing the shapefile and excel metadata file}

\item{file}{the shapefile name with extension shp}

\item{mfile}{the metadata file with an excel extension (e.g. xlsx or xls)}
}
\value{
a sf
}
\description{
Merge a compatible urban area sf with its matching metadata data frame
}
\examples{
#not run: `mergeUA("~", "sf_2000_reconcile.shp", "sf_2000_reconcile.xlsx")`
}
