% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{meta2000}
\alias{meta2000}
\title{US Census Bureau 2000 Census Urban Area Metadata data frame}
\format{
A data frame with 3637 rows and 5 variables:
\describe{
\item{Code}{the urban area UACE code}
\item{Name}{the urban area name}
\item{Population}{the urban area population}
\item{Area(sq meters)}{the area of the urban area}
\item{PopulationDensity}{the urban area population density per square mile}
}
}
\usage{
data("meta2000")
}
\description{
A data frame containing the 2000 US Census urban area metadata. Obtained from
https://www.census.gov/programs-surveys/geography/guidance/geo-areas/urban-rural/2000-urban-rural.html
on Aug 20, 2023. August 23, 2002 Changes Alphabetically-sorted corrected National list of UAs was
combined with November 20, 2002 Changes Alphabetically-sorted corrected National list of UCs.
The variables are as follows:
}
\keyword{datasets}
