% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_summaries.R
\name{metaUA}
\alias{metaUA}
\title{Read an urban area metadata file}
\usage{
metaUA(wd, mfile)
}
\arguments{
\item{wd}{a directory containing the excel metadata file}

\item{mfile}{the file name}
}
\value{
a data frame
}
\description{
Read an urban area metadata file
}
\examples{
#not run: `metaUA("~", "ua_comb00_natl_corr.xlsx")`
}
