% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_summaries.R
\name{shapeUA}
\alias{shapeUA}
\title{Read an urban area shapefile}
\usage{
shapeUA(wd, file)
}
\arguments{
\item{wd}{a directory containing the shapefile}

\item{file}{the file name}
}
\value{
a sf
}
\description{
Read an urban area shapefile
}
\examples{
#not run: `shapeUA("~", "ua99_d00.shp")`
}
