% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ua2000}
\alias{ua2000}
\title{US Census Bureau 2000 Census Urban Area Shapefile sf}
\format{
A sf with 11880 rows and 9 variables:
\describe{
\item{AREA}{AREA}
\item{PERIMETER}{PERIMETER}
\item{UA99_D00_}{UA99_D00_}
\item{UA99_D00_I}{UA99_D00_I}
\item{UA}{the urban area UACE code}
\item{NAME}{the urban area name}
\item{LSAD}{LSAD}
\item{LSAD_TRANS}{LSAD_TRANS}
\item{geometry}{the urban area feature geometry}
}
}
\usage{
data("ua2000")
}
\description{
A sf containing the 2000 US Census urban areas. Obtained from
https://www.census.gov/geographies/mapping-files/time-series/geo/carto-boundary-file.2000.html#list-tab-1556094155
on Aug 20, 2023. The variables are as follows:
}
\keyword{datasets}
