% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_summaries.R
\name{unionUA}
\alias{unionUA}
\title{Union a merged urban area sf (see ?mergeUA) when neighbor features are separated by less than a combine distance in miles}
\usage{
unionUA(ua, cdist)
}
\arguments{
\item{ua}{a merged sf}

\item{cdist}{a combine distance (in miles)}
}
\value{
a sf
}
\description{
Union a merged urban area sf (see ?mergeUA) when neighbor features are separated by less than a combine distance in miles
}
\examples{
#none
}
